Adds capability to attach other files on invoice
