+++ src/hocs/editInvoice.js
@@ -18,6 +18,7 @@ const editInvoiceConnector = connect(
     editedInvoiceToken: sel.editInvoiceToken,
     submitError: sel.apiEditInvoiceError,
     invoice: sel.invoice,
+    otherFiles: sel.getNotJSONFile,
     isLoading: or(sel.isLoadingSubmit, sel.isApiRequestingEditInvoice),
     loggedInAsEmail: sel.loggedInAsEmail,
     userCanExecuteActions: sel.userCanExecuteActions,
@@ -71,7 +72,8 @@ class EditInvoiceContainer extends Component {
           location: input.contractorlocation,
           rate: input.contractorrate,
           address: input.paymentaddress,
-          datasheet: convertLineItemsToGrid(input.lineitems, false)
+          datasheet: convertLineItemsToGrid(input.lineitems, false),
+          files: this.props.otherFiles
         }
       });
     }
