+++ src/lib/api.js
@@ -67,7 +67,8 @@ export const makeInvoice = (
   contractorcontact,
   contractorrate,
   paymentaddress,
-  lineItems
+  lineItems,
+  files = []
 ) => {
   const { name, mime, payload } = convertJsonToFile({
     version: 1,
@@ -95,8 +96,14 @@ export const makeInvoice = (
         mime,
         payload,
         digest: digestPayload(payload)
-      }
-    ]
+      },
+      ...files
+    ].map(({ name, mime, payload }) => ({
+      name,
+      mime,
+      payload,
+      digest: digestPayload(payload)
+    }))
   };
 };
